+++ tests/chainer_tests/dataset_tests/tabular_tests/test_as_mode.py
@@ -1,9 +1,9 @@
 import unittest
 
+import chainer
 from chainer import testing
-from chainer.dataset import TabularDataset
 
-from .test_tabular_dataset import DummyDataset
+from . import dummy_dataset
 
 
 @testing.parameterize(
@@ -13,9 +13,9 @@ from .test_tabular_dataset import DummyDataset
 class TestAsTuple(unittest.TestCase):
 
     def test_as_tuple(self):
-        dataset = DummyDataset(mode=self.mode)
+        dataset = dummy_dataset.DummyDataset(mode=self.mode)
         view = dataset.as_tuple()
-        self.assertIsInstance(view, TabularDataset)
+        self.assertIsInstance(view, chainer.dataset.TabularDataset)
         self.assertEqual(len(view), len(dataset))
         self.assertEqual(view.keys, dataset.keys)
         self.assertEqual(view.mode, tuple)
@@ -28,9 +28,9 @@ class TestAsTuple(unittest.TestCase):
 class TestAsDict(unittest.TestCase):
 
     def test_as_dict(self):
-        dataset = DummyDataset(mode=self.mode)
+        dataset = dummy_dataset.DummyDataset(mode=self.mode)
         view = dataset.as_dict()
-        self.assertIsInstance(view, TabularDataset)
+        self.assertIsInstance(view, chainer.dataset.TabularDataset)
         self.assertEqual(len(view), len(dataset))
         self.assertEqual(view.keys, dataset.keys)
         self.assertEqual(view.mode, dict)
